+++ pygametemplate/core.py
@@ -7,7 +7,7 @@ import ctypes
 
 import pygame
 
-from pygametemplate.exceptions import CaughtFatalException
+from pygametemplate.exceptions import CaughtFatalException, PygameError
 
 
 TEST = bool(int(os.environ.get("TEST", "0")))
@@ -16,7 +16,7 @@ TEST = bool(int(os.environ.get("TEST", "0")))
 PATH = os.getcwd()
 
 def path_to(*path):
-    """Returns the complete absolute path of the path given."""
+    """Return the complete absolute path of the path given."""
     return os.path.join(PATH, *"/".join(path).split("/"))
 
 
@@ -37,33 +37,38 @@ def log(*error_message, **options):
                 "Please check log.txt for details.").format(error_message)
         ctypes.windll.user32.MessageBoxW(0, text, "Error", 0)
         raise CaughtFatalException(sys.exc_info()[1])
-    else:
-        pass    # TODO: Add some code here to show an error message in game
+
+    # TODO: Add some code here to show an error message in game
 
 
 # Asset loading
-def load_image(image_name, fade_enabled=False, file_extension=".png"):
-    """fade_enabled should be True if you want images to be able to fade"""
+def load_image(image_name, fix_alphas=True, file_extension=".png"):
+    """Load the image with the given `image_name` (excluding file extension).
+
+    Setting `fix_alphas` to False enables the image to be able to fade.
+    A different file extension can be specified via the
+    `file_extension` keyword argument, which defaults to ".png".
+    """
+    # TODO: Add stuff for loading images of the correct resolution
+    # depending on the player's resolution settings.
+    image_path = path_to("assets/images", image_name + file_extension)
     try:
-        #! Add stuff for loading images of the correct resolution
-        # depending on the player's resolution settings
-        image_path = path_to("assets/images", image_name + file_extension)
         try:
             image = pygame.image.load(image_path)
-        except pygame.error:
+        except PygameError:
             raise IOError
-        if not fade_enabled:
-            return image.convert_alpha()  # Fixes per pixel alphas permanently
-        else:
-            return image.convert()
-    except Exception:
-        log("Failed to load image: ", image_name, file_extension)
+    except IOError:
+        log("Image file not found: ", image_name, file_extension)
+
+    if fix_alphas:
+        return image.convert_alpha()  # Fixes per pixel alphas permanently
+    return image.convert()
 
 
 def load_font(font_name, font_size, file_extension=".ttf"):
+    """Load the font with the given `font_name` with the given `font_size`."""
+    font_path = path_to("assets/fonts", font_name + file_extension)
     try:
-        return pygame.font.Font(
-            path_to("assets/fonts", font_name + file_extension), font_size
-        )
-    except Exception:
-        log("Failed to load font: ", font_name, file_extension)
+        return pygame.font.Font(font_path, font_size)
+    except IOError:
+        log("Font file not found: ", font_name, file_extension)