+++ pygametemplate/debug/__init__.py
@@ -1,0 +1,2 @@
+"""Module containing useful debugging tools for making games with pygametemplate."""
+from .debug_console import DebugConsole